+++ core/tests/bootstrap.php
@@ -71,6 +71,7 @@ namespace
 	require_once dirname(__DIR__) . '/auth/image_authorization.php';
 	require_once dirname(__DIR__) . '/ext.php';
 	require_once dirname(__DIR__) . '/controller/moderate.php';
+	require_once dirname(__DIR__) . '/controller/index.php';
 	require_once dirname(__DIR__) . '/acp/main_module.php';
 	require_once dirname(__DIR__) . '/ucp/main_module.php';
 	require_once dirname(__DIR__) . '/config.php';